refactor(aiida): InboundDataSource refactoring - Rename AbstractAggregator to Aggregator and the old Aggregator to OutboundAggregator - Rename AbstractDataRecord to DataSourceRecord - Consolidated Flux of InboundAdapter to use DataSourceAdapter's inherited one for all adapters to work similarly - Other small PR annotations
