+++ aiida/src/main/java/energy/eddie/aiida/aggregator/Aggregator.java
@@ -1,186 +1,109 @@
-// SPDX-FileCopyrightText: 2023-2026 The EDDIE Developers <eddie.developers@fh-hagenberg.at>
+// SPDX-FileCopyrightText: 2026 The EDDIE Developers <eddie.developers@fh-hagenberg.at>
 // SPDX-License-Identifier: Apache-2.0
 
 package energy.eddie.aiida.aggregator;
 
 import energy.eddie.aiida.adapters.datasource.DataSourceAdapter;
 import energy.eddie.aiida.models.datasource.DataSource;
-import energy.eddie.aiida.models.record.AiidaRecord;
-import energy.eddie.aiida.models.record.AiidaRecordValue;
-import energy.eddie.aiida.repositories.AiidaRecordRepository;
-import energy.eddie.api.agnostic.aiida.AiidaAsset;
-import energy.eddie.api.agnostic.aiida.ObisCode;
+import energy.eddie.aiida.models.record.DataSourceRecord;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.boot.health.registry.HealthContributorRegistry;
-import org.springframework.scheduling.concurrent.ThreadPoolTaskScheduler;
-import org.springframework.scheduling.support.CronExpression;
-import org.springframework.scheduling.support.CronTrigger;
-import org.springframework.stereotype.Component;
-import reactor.core.publisher.Flux;
 import reactor.core.publisher.Sinks;
+import reactor.core.scheduler.Schedulers;
 
-import java.time.Instant;
-import java.util.*;
-import java.util.function.Function;
-import java.util.stream.Collectors;
-import java.util.stream.Stream;
+import java.util.ArrayList;
+import java.util.List;
 
-import static java.util.Comparator.comparing;
-import static java.util.function.BinaryOperator.maxBy;
-
-@Component
-public class Aggregator extends AbstractAggregator<AiidaRecord> {
+public abstract class Aggregator<T extends DataSourceRecord> implements AutoCloseable {
     private static final Logger LOGGER = LoggerFactory.getLogger(Aggregator.class);
+    private static final String HEALTH_REGISTRY_PREFIX = "DATA_SOURCE_";
+    protected final List<DataSourceAdapter<? extends DataSource>> dataSourceAdapters;
+    protected final HealthContributorRegistry healthContributorRegistry;
+    private final Class<T> recordType;
+    protected Sinks.Many<DataSourceRecord> combinedRecordSink;
 
-    private final AiidaRecordRepository aiidaRecordRepository;
-
-    public Aggregator(
-            AiidaRecordRepository aiidaRecordRepository,
+    protected Aggregator(
+            Class<T> recordType,
             HealthContributorRegistry healthContributorRegistry
     ) {
-        super(healthContributorRegistry);
-        this.aiidaRecordRepository = aiidaRecordRepository;
+        this.recordType = recordType;
+        this.healthContributorRegistry = healthContributorRegistry;
+
+        combinedRecordSink = Sinks.many().multicast().directAllOrNothing();
+        combinedRecordSink.asFlux()
+                          .ofType(recordType)
+                          .publishOn(Schedulers.boundedElastic())
+                          .doOnNext(this::saveRecordToDatabase)
+                          .doOnError(this::handleCombinedSinkError)
+                          .subscribe();
+
+        dataSourceAdapters = new ArrayList<>();
     }
 
-    @Override
+    /**
+     * Adds a new {@link DataSourceAdapter} to this aggregator and will subscribe to the Flux returned by
+     * {@link DataSourceAdapter#start()}.
+     *
+     * @param dataSourceAdapter The new data source adapter to add. No check is made if this is a duplicate.
+     */
     public void addNewDataSourceAdapter(DataSourceAdapter<? extends DataSource> dataSourceAdapter) {
-        super.addNewDataSourceAdapter(dataSourceAdapter);
+        var dataSource = dataSourceAdapter.dataSource();
+        LOGGER.info("Will add datasource {} with ID {} to aggregator", dataSource.name(), dataSource.id());
+
+        healthContributorRegistry.registerContributor(HEALTH_REGISTRY_PREFIX + dataSource.id(), dataSourceAdapter);
+        dataSourceAdapters.add(dataSourceAdapter);
+
         dataSourceAdapter.start()
+                         .ofType(recordType)
                          .subscribe(this::publishRecordToCombinedFlux,
                                     throwable -> handleError(throwable, dataSourceAdapter));
     }
 
-    @Override
-    protected void saveRecordToDatabase(AiidaRecord dataRecord) {
-        LOGGER.trace("Saving new AIIDA record to db");
-        for (AiidaRecordValue value : dataRecord.aiidaRecordValues()) {
-            value.setAiidaRecord(dataRecord);
-        }
-        aiidaRecordRepository.save(dataRecord);
+    /**
+     * Removes a {@link DataSourceAdapter} from this aggregator and will close the datasource.
+     *
+     * @param dataSourceAdapter The datasource adapter to remove.
+     */
+    public void removeDataSourceAdapter(DataSourceAdapter<? extends DataSource> dataSourceAdapter) {
+        var dataSource = dataSourceAdapter.dataSource();
+        LOGGER.info("Will remove datasource {} with ID {} from aggregator", dataSource.name(), dataSource.id());
+
+        healthContributorRegistry.unregisterContributor(HEALTH_REGISTRY_PREFIX + dataSource.id());
+        dataSourceAdapters.remove(dataSourceAdapter);
+        dataSourceAdapter.close();
     }
 
     /**
-     * Returns a Flux of {@link AiidaRecord}s that either contains all records or only contains records with a {@link AiidaRecordValue#dataTag()}
-     * that is in the set {@code allowedCodes}.
-     * All values must have a timestamp before {@code permissionExpirationTime}.
-     * Additionally, the records are buffered and aggregated by the {@link CronExpression} {@code transmissionSchedule}.
-     *
-     * @param allowedDataTags          Tags which should be included in the returned Flux.
-     * @param allowedAsset             The asset that should be included in the returned Flux.
-     * @param permissionExpirationTime Instant when the permission expires.
-     * @param transmissionSchedule     The schedule at which the data should be transmitted.
-     * @param userId                   The user id of the permission creator.
-     * @return A Flux on which will only have AiidaRecords with a code matching one of the codes of the
-     * {@code allowedCodes} set, a timestamp that is before {@code permissionExpirationTime} and that are aggregated
-     * by the {@code transmissionSchedule}.
+     * Closes all {@link energy.eddie.aiida.adapters.datasource.inbound.InboundAdapter}s and emits a complete signal for all the Flux returned by
      */
-    @SuppressWarnings("FutureReturnValueIgnored")
-    public Flux<AiidaRecord> getFilteredFlux(
-            Set<ObisCode> allowedDataTags,
-            AiidaAsset allowedAsset,
-            Instant permissionExpirationTime,
-            CronExpression transmissionSchedule,
-            UUID userId,
-            UUID dataSourceId
-    ) {
-        var cronSink = Sinks.many().multicast().directAllOrNothing();
-        var cronTrigger = new CronTrigger(transmissionSchedule.toString());
-        var cronScheduler = new ThreadPoolTaskScheduler();
+    @Override
+    public void close() {
+        // ignore if complete signal can't be emitted successfully
+        combinedRecordSink.tryEmitComplete();
 
-        cronScheduler.initialize();
-        cronScheduler.schedule(() -> cronSink.tryEmitNext(true), cronTrigger);
-
-        var flux = combinedRecordSink.asFlux().doOnComplete(() -> {
-            cronScheduler.stop();
-            cronSink.tryEmitComplete();
-        });
-
-        return flux.map(AiidaRecord::new)
-                   .filter(aiidaRecord -> isValidAiidaRecord(aiidaRecord,
-                                                             allowedAsset,
-                                                             permissionExpirationTime,
-                                                             userId,
-                                                             dataSourceId))
-                   .map(aiidaRecord -> filterAllowedDataTags(aiidaRecord, allowedDataTags))
-                   .buffer(cronSink.asFlux())
-                   .flatMapIterable(this::aggregateRecords)
-                   .onErrorContinue((err, obj) -> LOGGER.error("Error while filtering records for permission", err));
-    }
-
-    private boolean isValidAiidaRecord(
-            AiidaRecord aiidaRecord,
-            AiidaAsset allowedAsset,
-            Instant expirationTime,
-            UUID userId,
-            UUID dataSourceId
-    ) {
-        var dataSource = aiidaRecord.dataSource();
-        return isAllowedAsset(dataSource, allowedAsset) &&
-               areAiidaRecordValuesValid(aiidaRecord.aiidaRecordValues()) &&
-               isBeforeExpiration(aiidaRecord, expirationTime) &&
-               isSameDataSource(dataSource, dataSourceId) &&
-               doesDataSourceBelongToCurrentUser(dataSource, userId);
-    }
-
-    private boolean isAllowedAsset(DataSource dataSource, AiidaAsset allowedAsset) {
-        return dataSource.asset() == allowedAsset;
-    }
-
-    private boolean areAiidaRecordValuesValid(List<AiidaRecordValue> aiidaRecordValues) {
-        return !aiidaRecordValues.isEmpty();
-    }
-
-    private boolean isSameDataSource(DataSource dataSource, UUID dataSourceId) {
-        return dataSource.id().equals(dataSourceId);
-    }
-
-    private boolean doesDataSourceBelongToCurrentUser(DataSource dataSource, UUID userId) {
-        return dataSource.userId().equals(userId);
-    }
-
-    private boolean isBeforeExpiration(AiidaRecord aiidaRecord, Instant permissionExpirationTime) {
-        return aiidaRecord.timestamp().isBefore(permissionExpirationTime);
-    }
-
-    private AiidaRecord filterAllowedDataTags(AiidaRecord aiidaRecord, Set<ObisCode> allowedDataTags) {
-        if (!allowedDataTags.isEmpty()) {
-            var filteredValues = aiidaRecord.aiidaRecordValues()
-                                            .stream()
-                                            .filter(value -> allowedDataTags.contains(value.dataTag()))
-                                            .toList();
-            aiidaRecord.setAiidaRecordValues(filteredValues);
+        LOGGER.info("Closing all {} datasources", dataSourceAdapters.size());
+        for (var dataSourceAdapter : dataSourceAdapters) {
+            dataSourceAdapter.close();
         }
-
-        return aiidaRecord;
     }
 
-    private List<AiidaRecord> aggregateRecords(List<AiidaRecord> aiidaRecords) {
-        var aggregatedRecords = aiidaRecords.stream()
-                                            .collect(Collectors.toMap(
-                                                    r -> r.dataSource().id(),
-                                                    Function.identity(),
-                                                    this::mergeRecords,
-                                                    LinkedHashMap::new
-                                            ));
-        return new ArrayList<>(aggregatedRecords.values());
+    protected abstract void saveRecordToDatabase(T dataRecord);
+
+    protected synchronized void publishRecordToCombinedFlux(T data) {
+        var result = combinedRecordSink.tryEmitNext(data);
+
+        if (result.isFailure()) {
+            LOGGER.error("Error while emitting record to combined sink. Error was: {}", result);
+        }
     }
 
-    private AiidaRecord mergeRecords(AiidaRecord r1, AiidaRecord r2) {
-        var latestRecord = r2.timestamp().isAfter(r1.timestamp()) ? r2 : r1;
-        Map<String, AiidaRecordValue> mergedValues = new HashMap<>();
+    protected void handleCombinedSinkError(Throwable throwable) {
+        LOGGER.error("Error from combined sink", throwable);
+    }
 
-        Stream.concat(r1.aiidaRecordValues().stream(), r2.aiidaRecordValues().stream())
-              .forEach(val -> mergedValues.merge(
-                      val.rawTag(),
-                      val,
-                      maxBy(comparing(v -> v.aiidaRecord().timestamp()))
-              ));
-
-        return new AiidaRecord(
-                latestRecord.timestamp(),
-                latestRecord.dataSource(),
-                new ArrayList<>(mergedValues.values())
-        );
+    protected void handleError(Throwable throwable, DataSourceAdapter<? extends DataSource> dataSourceAdapter) {
+        // TODO: GH-2503 do we try to restart the affected datasource or only notify user?
+        LOGGER.error("Error from datasource {}", dataSourceAdapter.dataSource().name(), throwable);
     }
 }
